LPS-101511 Apply AnalyticsClientRequestContributor to segments-experiment-web module
