+++ modules/apps/segments/segments-experiment-web/build.gradle
@@ -11,6 +11,7 @@ dependencies {
 	compileOnly group: "javax.servlet.jsp", name: "javax.servlet.jsp-api", version: "2.3.1"
 	compileOnly group: "org.apache.felix", name: "org.apache.felix.http.servlet-api", version: "1.1.2"
 	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
+	compileOnly project(":apps:analytics:analytics-api")
 	compileOnly project(":apps:frontend-taglib:frontend-taglib")
 	compileOnly project(":apps:layout:layout-content-page-editor-api")
 	compileOnly project(":apps:product-navigation:product-navigation-control-menu-api")
