+++ modules/apps/segments/segments-experiment-web/src/main/java/com/liferay/segments/experiment/web/internal/constants/SegmentsExperimentWebKeys.java
@@ -27,7 +27,4 @@ public class SegmentsExperimentWebKeys {
 	public static final String SEGMENTS_EXPERIMENT_PANEL_STATE_OPEN =
 		"SEGMENTS_EXPERIMENT_PANEL_STATE_OPEN";
 
-	public static final String SEGMENTS_EXPERIMENT_SEGMENTS_EXPERIENCE_KEY =
-		"SEGMENTS_EXPERIMENT_SEGMENTS_EXPERIENCE_KEY";
-
 }
