+++ modules/apps/segments/segments-experiment-web/src/main/java/com/liferay/segments/experiment/web/internal/servlet/taglib/SegmentsExperimentAnalyticsTopHeadJSPDynamicInclude.java
@@ -19,14 +19,8 @@ import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.servlet.taglib.BaseJSPDynamicInclude;
 import com.liferay.portal.kernel.servlet.taglib.DynamicInclude;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.WebKeys;
-import com.liferay.segments.constants.SegmentsExperienceConstants;
-import com.liferay.segments.constants.SegmentsWebKeys;
-import com.liferay.segments.experiment.web.internal.constants.SegmentsExperimentWebKeys;
 import com.liferay.segments.experiment.web.internal.util.SegmentsExperimentUtil;
-import com.liferay.segments.model.SegmentsExperience;
-import com.liferay.segments.service.SegmentsExperienceLocalService;
 
 import java.io.IOException;
 
@@ -60,15 +54,6 @@ public class SegmentsExperimentAnalyticsTopHeadJSPDynamicInclude
 			return;
 		}
 
-		long[] segmentsExperienceIds = GetterUtil.getLongValues(
-			httpServletRequest.getAttribute(
-				SegmentsWebKeys.SEGMENTS_EXPERIENCE_IDS));
-
-		httpServletRequest.setAttribute(
-			SegmentsExperimentWebKeys.
-				SEGMENTS_EXPERIMENT_SEGMENTS_EXPERIENCE_KEY,
-			_getSegmentsExperienceKey(segmentsExperienceIds));
-
 		super.include(httpServletRequest, httpServletResponse, key);
 	}
 
@@ -97,24 +82,7 @@ public class SegmentsExperimentAnalyticsTopHeadJSPDynamicInclude
 		super.setServletContext(servletContext);
 	}
 
-	private String _getSegmentsExperienceKey(long[] segmentsExperienceIds) {
-		if (segmentsExperienceIds.length > 0) {
-			SegmentsExperience segmentsExperience =
-				_segmentsExperienceLocalService.fetchSegmentsExperience(
-					segmentsExperienceIds[0]);
-
-			if (segmentsExperience != null) {
-				return segmentsExperience.getSegmentsExperienceKey();
-			}
-		}
-
-		return SegmentsExperienceConstants.KEY_DEFAULT;
-	}
-
 	private static final Log _log = LogFactoryUtil.getLog(
 		SegmentsExperimentAnalyticsTopHeadJSPDynamicInclude.class);
-
-	@Reference
-	private SegmentsExperienceLocalService _segmentsExperienceLocalService;
 
 }
