+++ modules/apps/segments/segments-experiment-web/src/main/resources/META-INF/resources/init.jsp
@@ -29,8 +29,7 @@ page import="com.liferay.portal.kernel.util.GetterUtil" %><%@
 page import="com.liferay.portal.kernel.util.PortalUtil" %><%@
 page import="com.liferay.segments.experiment.web.internal.constants.SegmentsExperimentWebKeys" %><%@
 page import="com.liferay.segments.experiment.web.internal.display.context.SegmentsExperimentDisplayContext" %><%@
-page import="com.liferay.segments.experiment.web.internal.util.SegmentsExperimentUtil" %><%@
-page import="com.liferay.segments.model.SegmentsExperiment" %>
+page import="com.liferay.segments.experiment.web.internal.util.SegmentsExperimentUtil" %>
 
 <liferay-frontend:defineObjects />
